all managements completed succesfully

+++ ProductApp/views.py
@@ -6,7 +6,7 @@ from .serializer import ProductSerializer
 from rest_framework import status
 from .models import Products
 
-
+# product management view
 class ProductManagment(APIView):
     permission_classes = [IsAuthenticated]
     # adding the data to the database
@@ -33,23 +33,24 @@ class ProductManagment(APIView):
             return Response(serializer.data,status=status.HTTP_302_FOUND)
     
     # updating the details
-    def patch(self,request,item_id):
-        print(item_id)
+    def patch(self,request, item_id=None):
+        if item_id is None:
+            return Response({'error': 'Item ID is required.'}, status=status.HTTP_400_BAD_REQUEST)
         try:
             item = Products.objects.get(id = item_id)
         except:
             return Response({'error':"Item is not found"},status=status.HTTP_404_NOT_FOUND)
-        print(item)
         serializer = ProductSerializer(item,data = request.data, partial=True)
         if serializer.is_valid():
-            print(serializer.validated_data)
             serializer.save()
             return Response(serializer.data,status=status.HTTP_200_OK)
         return Response(serializer.errors,status=status.HTTP_400_BAD_REQUEST)
         
         
     # deleting the product
-    def delete(self,request,item_id):
+    def delete(self,request, item_id=None):
+        if item_id is None:
+            return Response({'error': 'Item ID is required.'}, status=status.HTTP_400_BAD_REQUEST)
         try:
             item = Products.objects.get(id = item_id)
             item.delete()
